IDO: Properly cache notification_id for contactnotifications
1) We cannot "pass" the notification insert id directly
to the following contactnotifications queries. Instead
we need to properly initialize the cache and later fetch
the correct notification_id
2) There's the possibility that 2+ notifications happen
for the same notification. In that case the old notification_id
cache is invalid as it does not respect the timestamp. This patch
adds such a cache.

3) We need to clean unused notification_id entries
from the cache. Otherwise we would leak memory holding references
to the object pointer and timestamp.

refs #11387